通知方法更新

+++ sunny/Lib/ReachabilityCheck/ReachabilityCheck.m
@@ -29,7 +29,7 @@
             return YES;
             break;
         default:
-            NSLog(@"？？[%d]", status);
+            NSLog(@"？？[%ld]", (long)status);
             return NO;
             break;
     }
